backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# Environment
python-dotenv==1.0.0

# AI & LangChain
langchain==0.3.0
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Image Processing
pillow==11.0.0

# PDF Processing
PyPDF2==3.0.1
